Move test setup code into common module
* Move global dispatch test into integration tests
* Use common subscribers for dispatch tests
